refactor: satisfy the complexity ratchet on the merged tree
The rebase onto PR-A unioned two independently-green branches, so four
baseline-locked files grew past their owner-locked caps. Split each to a
smaller owner instead of raising the bar:

- packages/orchestrator/src/orchestrator.ts: the credential-resolution
  cluster (strict-pin resolver, paid-route flags, A7-aware readiness epoch,
  differential-probe wiring, the preflight wrapper over account-resolution)
  moves to orchestrator-credentials.ts behind an accessor-function host.
- packages/daemon/src/quota-registry.ts: the pure per-snapshot helpers
  (v3.2.0 rollback shaping, snapshot keying, freshness aging, expired
  scoped-cooldown pruning) move to quota-registry-support.ts.
- apps/macos AccountsPopover.swift: the tri-state auto-switch control moves
  to AccountsAutoBalanceControl.swift (pure control surface; the target-set
  logic stays in AccountsPresentation.swift).
- apps/macos AuthSheet.swift: the native-setup and API-key panels move to
  AuthSheetNativeSetupPanel.swift (pure rendering; the sheet keeps every
  lifecycle mutation), and the close-confirmation copy joins
  AuthSheetClosePolicy in AuthSheetPresentation.swift. AppModel.swift
  comment tightening covers the two-line union overhang.

No behavior change; Kit 289 + App 685 Swift tests and the node typechecks
pass on the split tree.
